+++ packages/terra-dropdown-button/src/SplitButton.jsx
@@ -49,6 +49,16 @@ const propTypes = {
    * The intl object to be injected for translations.
    */
   intl: intlShape.isRequired,
+  /**
+   * @private
+   * Callback to tell the parent it should close the dropdown
+   */
+  requestClose: PropTypes.func,
+  /**
+   * The associated metaData to be provided in the onSelect callback.
+   */
+  // eslint-disable-next-line react/forbid-prop-types
+  metaData: PropTypes.object,
 };
 
 const defaultProps = {
@@ -63,21 +73,61 @@ class SplitButton extends React.Component {
     super(props);
 
     this.handleDropdownButtonClick = this.handleDropdownButtonClick.bind(this);
+    this.handlePrimaryButtonClick = this.handlePrimaryButtonClick.bind(this);
     this.handleDropdownRequestClose = this.handleDropdownRequestClose.bind(this);
     this.handlePrimaryKeyDown = this.handlePrimaryKeyDown.bind(this);
     this.handlePrimaryKeyUp = this.handlePrimaryKeyUp.bind(this);
     this.handleCaretKeyDown = this.handleCaretKeyDown.bind(this);
     this.handleCaretKeyUp = this.handleCaretKeyUp.bind(this);
+    this.setButtonNode = this.setButtonNode.bind(this);
+    this.getButtonNode = this.getButtonNode.bind(this);
+    this.setListNode = this.setListNode.bind(this);
+    this.openDropDown = this.openDropDown.bind(this);
 
-    this.state = { isOpen: false, caretIsActive: false, primaryIsActive: false };
+    this.state = {
+      isOpen: false, caretIsActive: false, primaryIsActive: false, openedViaKeyboard: false,
+    };
   }
 
-  handleDropdownButtonClick() {
+  setListNode(element) {
+    this.dropdownList = element;
+  }
+
+  setButtonNode(node) {
+    this.buttonNode = node;
+  }
+
+  getButtonNode() {
+    return this.buttonNode;
+  }
+
+  openDropDown(event) {
     this.setState(prevState => ({ isOpen: !prevState.isOpen }));
+    // See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Button#Clicking_and_focus
+    // Button on Firefox, Safari and IE running on OS X does not receive focus when clicked.
+    // This will put focus on the button when clicked.
+    event.currentTarget.focus();
+  }
+
+  handleDropdownButtonClick(event) {
+    if (this.state.isOpen) {
+      this.setState({ openedViaKeyboard: false });
+    }
+    this.openDropDown(event);
+  }
+
+  handlePrimaryButtonClick(event) {
+    // See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Button#Clicking_and_focus
+    // Button on Firefox, Safari and IE running on OS X does not receive focus when clicked.
+    // This will put focus on the button when clicked.
+    event.currentTarget.focus();
+    this.props.onSelect(event, this.props.metaData);
+    this.handleDropdownRequestClose();
   }
 
   handleDropdownRequestClose(callback) {
-    this.setState({ isOpen: false }, typeof callback === 'function' ? callback : undefined);
+    const onSelectCallback = typeof callback === 'function' ? callback : undefined;
+    this.setState({ isOpen: false, openedViaKeyboard: false, caretIsActive: false }, onSelectCallback);
   }
 
   /*
@@ -96,14 +146,36 @@ class SplitButton extends React.Component {
   }
 
   handleCaretKeyDown(event) {
-    if (event.keyCode === KeyCode.KEY_SPACE) {
-      this.setState({ caretIsActive: true });
+    if (event.keyCode === KeyCode.KEY_SPACE || event.keyCode === KeyCode.KEY_RETURN) {
+      // In FireFox active styles don't get applied onKeyDown
+      this.setState({ caretIsActive: true, openedViaKeyboard: true });
+      /*
+        Prevent the callback from being called repeatedly if the RETURN or SPACE key is held down.
+        The keyDown event of native html button triggers Onclick() event on RETURN or SPACE key press.
+        where holding RETURN key for longer time will call dropdownClick() event repeatedly which would cause
+        the dropdown to open and close itself.
+      */
+      event.preventDefault();
+    } else if (event.keyCode === KeyCode.KEY_DOWN && this.state.isOpen && !this.state.openedViaKeyboard) {
+      // set focus to first list element on down arrow key press only when dropdown is opened by mouse click.
+      const listOptions = this.dropdownList.querySelectorAll('[data-terra-dropdown-list-item]');
+      listOptions[0].focus();
+      // prevent handleFocus() callback of DropdownList.
+      event.preventDefault();
+    } else if (event.keyCode === KeyCode.KEY_UP && this.state.isOpen && !this.state.openedViaKeyboard) {
+      // set focus to last list element on up arrow key press only when dropdown is opened by mouse click
+      const listOptions = this.dropdownList.querySelectorAll('[data-terra-dropdown-list-item]');
+      listOptions[listOptions.length - 1].focus();
+      event.preventDefault();
+    } else if (event.keyCode === KeyCode.KEY_TAB) {
+      this.handleDropdownRequestClose();
     }
   }
 
   handleCaretKeyUp(event) {
-    if (event.keyCode === KeyCode.KEY_SPACE) {
+    if (event.keyCode === KeyCode.KEY_SPACE || event.keyCode === KeyCode.KEY_RETURN) {
       this.setState({ caretIsActive: false });
+      this.openDropDown(event);
     }
   }
 
@@ -117,6 +189,8 @@ class SplitButton extends React.Component {
       onSelect,
       variant,
       intl,
+      requestClose,
+      metaData,
       ...customProps
     } = this.props;
 
@@ -124,6 +198,7 @@ class SplitButton extends React.Component {
       isOpen,
       primaryIsActive,
       caretIsActive,
+      openedViaKeyboard,
     } = this.state;
 
     const caretLabel = intl.formatMessage({ id: 'Terra.dropdownButton.moreOptions' });
@@ -155,11 +230,14 @@ class SplitButton extends React.Component {
         isBlock={isBlock}
         isCompact={isCompact}
         isDisabled={isDisabled}
+        openedViaKeyboard={openedViaKeyboard}
+        buttonRef={this.getButtonNode}
+        refCallback={this.setListNode}
       >
         <button
           type="button"
           className={primaryClassnames}
-          onClick={onSelect}
+          onClick={this.handlePrimaryButtonClick}
           onKeyDown={this.handlePrimaryKeyDown}
           onKeyUp={this.handlePrimaryKeyUp}
           disabled={isDisabled}
@@ -177,9 +255,10 @@ class SplitButton extends React.Component {
           disabled={isDisabled}
           tabIndex={isDisabled ? '-1' : undefined}
           aria-disabled={isDisabled}
-          aria-expanded={isOpen || undefined}
+          aria-expanded={isOpen}
           aria-haspopup="menu"
           aria-label={caretLabel}
+          ref={this.setButtonNode}
         >
           <span className={cx('caret-icon')} />
         </button>